refactor(world): extract useWorldMutations hook + doc drift fixes
Pulls the 10 useMutation(...).withOptimisticUpdate(...) declarations out
of src/routes/world.tsx (lines 96-282) into src/hooks/useWorldMutations.ts.
world.tsx drops from 900 to 715 lines; the hook takes movementSpeed as a
value and owns the ref internally, eliminating the stats/mutation
ordering dance that the old comment block had to explain.

The Convex mutation name `useTeleportStone` tripped biome's
useHookAtTopLevel rule when called from async handlers, since biome's
hook-detection only sees the local identifier. Aliased to
`teleportStone` in the destructure to keep the lint clean without a
biome-ignore suppression.

Docs drift caught alongside:
- codebase-map.md still said convex/characters.ts was a "kitchen sink
  pending split", but the split shipped earlier — characters.ts is now
  176 lines (CRUD only) and combat.ts / items.ts / vendor.ts own the
  rest. Replaced the single stale row with accurate per-file rows.
- codebase-map.md said world.tsx was a "mid-size file (~470 lines)" —
  it's 900, refactor in flight (this branch).
- in-progress.md gained a "Split useCombatLoop.ts (queued after world)"
  entry — at 916 lines it's the single largest file in the repo and
  becomes the new MODIFYING-friction worst case once world.tsx lands.
